requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
supabase==2.3.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
APScheduler==3.10.4
gunicorn==21.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
cryptography==41.0.7
resend==0.8.0
